requirements.txt
-e .  # Get self and dependencies from setup.py

    attrs
